Please enter the commit message for your changes. Lines starting

+++ ruby.cpp
@@ -89,15 +89,15 @@ int main()
                 get_pixel(sample.y, sample.x, 0),
                 (double)sample.r /((get_pixel(sample.y, sample.x, 1) + get_pixel(sample.y, sample.x, 2) + redThreshold )/ 2)};
 
-                if (sample.relativeRedBuffer1 != 1298)
+                if ((int)sample.relativeRedBuffer1 != 1298)
                 {
                     errors += (sample.relativeRed < sample.relativeRedBuffer1 - marginOfError ||
                     sample.relativeRed > sample.relativeRedBuffer1 + marginOfError ) ? 1 : 0;
 
 
                 }
-                                    std::cout << "buffrr = " + std::to_string(sample.relativeRedBuffer1);
-                    std::cout << " | relred = " + std::to_string(sample.relativeRed) << std::endl;
+                std::cout << "buffrr = " + std::to_string(sample.relativeRedBuffer1);
+                std::cout << " | relred = " + std::to_string(sample.relativeRed) << std::endl;
                 sample.relativeRedBuffer1 = sample.relativeRed;
                 set_pixel(sample.y, sample.x, sample.relativeRed,sample.relativeRed,sample.relativeRed);
             }
